requirements.txt
googletrans==3.1.0a0
matplotlib==3.6.2
nltk==3.8
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
seaborn==0.12.2
streamlit==1.16.0
textblob==0.17.1
tweepy==4.12.1
wordcloud==1.8.2.2
